requirements.txt
astroid==2.4.2
autopep8==1.5.3
bcrypt==3.1.7
Bootstrap-Flask==1.3.2
cachelib==0.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Session==0.3.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
